Move getStats to a dedicated file
