+++ lib/routes.js
@@ -16,7 +16,6 @@ import {
   decoratePointPrelevement,
   getPointPrelevement,
   getPointsFromPreleveur,
-  getStats,
   createPointPrelevement,
   updatePointPrelevement,
   deletePointPrelevement,
@@ -63,6 +62,8 @@ import {
   getMeso,
   getMesoList
 } from './models/referentiels.js'
+
+import {getStats} from './models/stats.js'
 
 const demarcheNumber = Number.parseInt(process.env.DS_DEMARCHE_NUMBER, 10)
